pyproject.toml
[tool.poetry]
name = "link-storage"
version = "0.1.0"
description = ""
authors = ["Yuhen Liubimov <yauhenliubimau@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
django = "^5.0.4"
djangorestframework = "^3.15.1"
environs = "^11.0.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
types-requests = "^2.31.0.20240406"
gunicorn = "^22.0.0"
drf-yasg = "^1.21.7"
setuptools = "^69.5.1"
pytest-django = "^4.8.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
mypy = "^1.9.0"
isort = "^5.13.2"


[tool.poetry.group.test.dependencies]
faker = "^24.11.0"

[tool.ruff]
cache-dir = ".venv/ruff/"
exclude = [
    ".git",
    ".git-rewrite",
    ".pyenv",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "site-packages",
    "postgres",
]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.mypy]
cache_dir = ".venv/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = true
exclude = [
    ".venv",
    ".github",
    ".idea",
    ".vscode",
    "venv",
    "postgres",
]
files = ["."]
pretty = true
python_version = "3.12"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.isort]
force_single_line = true
line_length = 79
multi_line_output = 3
profile = "black"
skip = [
    ".venv",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
cache_dir = ".venv/pytest/"
console_output_style = "count"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
